Add TaskCreateElement to AddTask
